Fix bug trying to use float as index for an array. Convert to int!

+++ midi_utility.py
@@ -425,8 +425,8 @@ def stylify_track(track, ticks_per_quarter, velocity_array, quantization):
         if hasattr(time_msg, 'time'):
             if time_msg.type == 'note_on' or time_msg.type == 'note_off':
                 if time_msg.velocity > 0:
-                    pos = cum_times[cum_index] * \
-                        (2**quantization/4) / (ticks_per_quarter)
+                    pos = int(cum_times[cum_index] * \
+                        (2**quantization/4) / (ticks_per_quarter))
                     if pos == normalized_num_steps:
                         pos = pos - 1
                     if pos > normalized_num_steps:
